requirements.txt
Werkzeug>=2.3.3
gunicorn>=20.1.0
cairosvg>=2.7.0
Flask>=2.3.2
opencv-python-headless>=4.8.1.78
numpy>=1.26.0
Pillow>=10.1.0
matplotlib>=3.7.2
scipy>=1.11.1
qrcode>=7.4.2